requirements.txt
ag2==0.9.7
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
asyncer==0.0.8
autogen==0.9.7
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.5
curl_cffi==0.12.0
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
fonttools==4.59.0
frozendict==2.4.6
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
investpy==1.0.8
jiter==0.10.0
kiwisolver==1.4.8
lxml==6.0.0
matplotlib==3.10.5
multitasking==0.0.12
numpy==2.3.2
openai==1.98.0
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
pip_system_certs==5.2
platformdirs==4.3.8
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydash==8.0.5
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
regex==2025.7.34
requests==2.32.4
requests-futures==1.0.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
termcolor==3.1.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
verify==1.1.1
webencodings==0.5.1
websockets==15.0.1
yahooquery==2.4.1
yfinance==0.2.36